Правки формата 4

+++ tasks/mpi/Sdobnov_V_sum_of_vector_elements/func_tests/main.cpp
@@ -21,7 +21,7 @@ TEST(Sdobnov_V_sum_of_vector_elements_par, EmptyOutput) {
   if (world.rank() == 0) {
     taskDataPar->inputs_count.emplace_back(rows);
     taskDataPar->inputs_count.emplace_back(columns);
-    for (int i = 0; i < input.size(); i++) {
+    for (long unsigned int i = 0; i < input.size(); i++) {
       taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t*>(input[i].data()));
     }
   }
@@ -41,7 +41,7 @@ TEST(Sdobnov_V_sum_of_vector_elements_par, EmptyMatrix) {
   if (world.rank() == 0) {
     taskDataPar->inputs_count.emplace_back(rows);
     taskDataPar->inputs_count.emplace_back(columns);
-    for (int i = 0; i < input.size(); i++) {
+    for (long unsigned int i = 0; i < input.size(); i++) {
       taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t*>(input[i].data()));
     }
     taskDataPar->outputs_count.emplace_back(1);
@@ -69,7 +69,7 @@ TEST(Sdobnov_V_sum_of_vector_elements_par, Matrix10x10) {
   if (world.rank() == 0) {
     taskDataPar->inputs_count.emplace_back(rows);
     taskDataPar->inputs_count.emplace_back(columns);
-    for (int i = 0; i < input.size(); i++) {
+    for (long unsigned int i = 0; i < input.size(); i++) {
       taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t*>(input[i].data()));
     }
     taskDataPar->outputs_count.emplace_back(1);
@@ -105,7 +105,7 @@ TEST(Sdobnov_V_sum_of_vector_elements_par, Matrix100x100) {
   if (world.rank() == 0) {
     taskDataPar->inputs_count.emplace_back(rows);
     taskDataPar->inputs_count.emplace_back(columns);
-    for (int i = 0; i < input.size(); i++) {
+    for (long unsigned int i = 0; i < input.size(); i++) {
       taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t*>(input[i].data()));
     }
     taskDataPar->outputs_count.emplace_back(1);
@@ -141,7 +141,7 @@ TEST(Sdobnov_V_sum_of_vector_elements_par, Matrix100x10) {
   if (world.rank() == 0) {
     taskDataPar->inputs_count.emplace_back(rows);
     taskDataPar->inputs_count.emplace_back(columns);
-    for (int i = 0; i < input.size(); i++) {
+    for (long unsigned int i = 0; i < input.size(); i++) {
       taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t*>(input[i].data()));
     }
     taskDataPar->outputs_count.emplace_back(1);
@@ -177,7 +177,7 @@ TEST(Sdobnov_V_sum_of_vector_elements_par, Matrix10x100) {
   if (world.rank() == 0) {
     taskDataPar->inputs_count.emplace_back(rows);
     taskDataPar->inputs_count.emplace_back(columns);
-    for (int i = 0; i < input.size(); i++) {
+    for (long unsigned int i = 0; i < input.size(); i++) {
       taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t*>(input[i].data()));
     }
     taskDataPar->outputs_count.emplace_back(1);
