+++ tasks/seq/Sdobnov_V_sum_of_vector_elements/func_tests/main.cpp
@@ -20,7 +20,7 @@ TEST(Sdobnov_V_sum_of_vector_elements_seq, EmptyOutput) {
 
   taskDataPar->inputs_count.emplace_back(rows);
   taskDataPar->inputs_count.emplace_back(columns);
-  for (int i = 0; i < input.size(); i++) {
+  for (long unsigned int i = 0; i < input.size(); i++) {
     taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t*>(input[i].data()));
   }
 
@@ -37,7 +37,7 @@ TEST(Sdobnov_V_sum_of_vector_elements_seq, EmptyMatrix) {
 
   taskDataPar->inputs_count.emplace_back(rows);
   taskDataPar->inputs_count.emplace_back(columns);
-  for (int i = 0; i < input.size(); i++) {
+  for (long unsigned int i = 0; i < input.size(); i++) {
     taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t*>(input[i].data()));
   }
   taskDataPar->outputs_count.emplace_back(1);
@@ -67,7 +67,7 @@ TEST(Sdobnov_V_sum_of_vector_elements_seq, Matrix10x10) {
 
   taskDataPar->inputs_count.emplace_back(rows);
   taskDataPar->inputs_count.emplace_back(columns);
-  for (int i = 0; i < input.size(); i++) {
+  for (long unsigned int i = 0; i < input.size(); i++) {
     taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t*>(input[i].data()));
   }
   taskDataPar->outputs_count.emplace_back(1);
@@ -97,7 +97,7 @@ TEST(Sdobnov_V_sum_of_vector_elements_seq, Matrix100x100) {
 
   taskDataPar->inputs_count.emplace_back(rows);
   taskDataPar->inputs_count.emplace_back(columns);
-  for (int i = 0; i < input.size(); i++) {
+  for (long unsigned int i = 0; i < input.size(); i++) {
     taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t*>(input[i].data()));
   }
   taskDataPar->outputs_count.emplace_back(1);
@@ -127,7 +127,7 @@ TEST(Sdobnov_V_sum_of_vector_elements_seq, Matrix100x10) {
 
   taskDataPar->inputs_count.emplace_back(rows);
   taskDataPar->inputs_count.emplace_back(columns);
-  for (int i = 0; i < input.size(); i++) {
+  for (long unsigned int i = 0; i < input.size(); i++) {
     taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t*>(input[i].data()));
   }
   taskDataPar->outputs_count.emplace_back(1);
@@ -157,7 +157,7 @@ TEST(Sdobnov_V_sum_of_vector_elements_seq, Matrix10x100) {
 
   taskDataPar->inputs_count.emplace_back(rows);
   taskDataPar->inputs_count.emplace_back(columns);
-  for (int i = 0; i < input.size(); i++) {
+  for (long unsigned int i = 0; i < input.size(); i++) {
     taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t*>(input[i].data()));
   }
   taskDataPar->outputs_count.emplace_back(1);
