+++ tasks/seq/Sdobnov_V_sum_of_vector_elements/perf_tests/main.cpp
@@ -21,7 +21,7 @@ TEST(Sdobnov_V_sum_of_vector_elements_seq, test_pipeline_run) {
 
   taskDataPar->inputs_count.emplace_back(rows);
   taskDataPar->inputs_count.emplace_back(columns);
-  for (int i = 0; i < input.size(); i++) {
+  for (long unsigned int i = 0; i < input.size(); i++) {
     taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t *>(input[i].data()));
   }
   taskDataPar->outputs_count.emplace_back(1);
@@ -65,7 +65,7 @@ TEST(Sdobnov_V_sum_of_vector_elements_seq, test_task_run) {
 
   taskDataPar->inputs_count.emplace_back(rows);
   taskDataPar->inputs_count.emplace_back(columns);
-  for (int i = 0; i < input.size(); i++) {
+  for (long unsigned int i = 0; i < input.size(); i++) {
     taskDataPar->inputs.emplace_back(reinterpret_cast<uint8_t *>(input[i].data()));
   }
   taskDataPar->outputs_count.emplace_back(1);
